requirements.txt
google-generativeai
streamlit
numpy<2.0.0
pandas
google-api-python-client
python-dotenv
pillow
plotly
streamlit-agraph
tinydb
rich
questionary
beautifulsoup4
requests
pypdf
openpyxl
# Black Ops Modules
playwright
stem
deepface
opencv-python
tf-keras
